implementado soma valor fracionado

+++ src/main/java/edu/fhg/tdd/CalculadoraTestada.java
@@ -1,7 +1,7 @@
 package edu.fhg.tdd;
 
 public class CalculadoraTestada {
-    public int somar(int valorA, int valorB) {
+    public double somar(double valorA, double valorB) {
         return valorA + valorB;
     }
 }
